+++ test/unit/cache/triggers/commutative/where_a_equal_any_b/where_cache_table_id_by_trigger_table_ids_with_cast/public.orders.sql
@@ -2,9 +2,9 @@ create or replace function cache_totals_for_companies_on_orders()
 returns trigger as $body$
 declare matched_old boolean;
 declare matched_new boolean;
-declare inserted_companies_ids integer[];
-declare not_changed_companies_ids integer[];
-declare deleted_companies_ids integer[];
+declare inserted_companies_ids bigint[];
+declare not_changed_companies_ids bigint[];
+declare deleted_companies_ids bigint[];
 begin
 
     if TG_OP = 'DELETE' then
